Finish second and final Purgatory level.
This commit wraps up the second and final Purgatory level. Some
interesting challenges await our hero, both from monsters and terrain.
This level also introduces the first encounter with a celestial dragon.
The 'boss' monster on this level is an archangel that guards the portal
exiting Purgatory. The exit deposits the player onto the beginning
sokoban level, and from there they can finish their ascension run up to
the Planes. The player can return to Purgatory if they want to, but
they'd have to travel all the way back down through gehennom to the
sanctum and use the entrance portal there. Or they can branchport back
to it if they have the Eye of the Aethiopica.

There's a couple other minor little tweaks here, including making sure
that archangels are geared out benefitting their station.