Forging artifacts.
With this commit, it is now possible to take two artifacts, and forge
them into a completely new artifact. Currently, there's just one recipe,
but more will be added soon in a future commit. Found and fixed a couple
bugs with artifact damage also while working on this. The breakdown:

* forging fire brand and frost brand together creates a brand new
artifact - the Sword of Annihilation
* the Sword of Annihilation is a steel long sword that has +d5 to-hit
and +d12 to damage, grants disintegration resistance when wielded, and
can disintegrate whatever monster (or player) it hits
* this new artifact is indestructible, does not appear randomly, and
cannot be wished for. It can only come into existience via forging
* artifacts cannot be forged with regular objects
* once an artifact is forged, the forge used will cool down and can no
longer be used

The bugs I noticed and fixed - special damage being applied to a monster
that was already set to be killed outright (would cause that monster to
be killed twice if they were wearing an amulet of life saving), and a
couple spots where a dmonsfree could occur if the player were subject to
a couple type of artifact instadeaths while wearing an amulet of life
saving.

There's more to do here. I'd like to play around with the weapon
feedback for the new artifact. And of course create a few more recipes
for artifacts. But it's late, so we'll save that for another time.

Thanks qt for the assist with the artifact forging bit.

Co-authored-by: entrez <me@entrez.cc>