New forging artifact recipe - Glamdring.
The second artifact recipe from forging - Glamdring. This was the elven
sword used by Gandalf during many of his travels and adventures. Just as
with any other forged artifact, it will not spawn randomly, and cannot
be wished for. To create it, Orcrist and Sting must be forged together.

The specs on this elven long sword aren't very epic, but considering
that Orcrist and Sting are the two easiest artifacts in the game to
obtain, this could be forged fairly early on, and it would definitely
make a difference for our hero in the early game. It's stats: Glamdring
is an elven long sword made of mithril. It has a +d8 to-hit and does
double damage to orcs and the like. Like its brethren, it will detect
orcs and glow blue in their presence. It confers shock resistance and
MC1 when wielded. It is chaotic for alignment purposes.

In this commit, I also updated the Blacksmith's Cookbook, it hints of
artifact forging, but does not list the recipes (some things the player
just needs to discover through trial and error). A livelog event has
also been created for foring an artifact.

If a wizard forges Glamdring, long sword skill is unlocked for them, and
can be advanced to skilled. Have fun, Gandalf.