New forging artifact recipe - The Staff of the Archmagi.
We've made enough long swords, time to make something for the primary
spellcasting types - The Staff of the Archmagi. This is also a forged
artifact, and cannot be obtained by any other means. It can be created
by forging Magicbane and Vorpal Blade together. The stats:

* appears as a white 'ashwood staff', base object is a quaterstaff, but
with a slightly higher damage modifer vs other quaterstaves
* when wielded, it reduces spell failure chances by 50% across the
board, all spell schools
* it also gives off light when wielded, BUC status affecting the output
radius
* confers magic resistance when wielded, half spell damage with carried
* damage type is the same as that as Magicbane, minus the cancellation
attack. to-hit/damage is +d5/+d8
* can be invoked to temporarily phase through solid matter
* is neutral in alignment

This is a fairly powerful artifact, but you're sacrificing two powerful
artifacts to make it, so I feel the trade-off is fair. I may play around
with one a bit more, we'll see.

Also in this commit, some bits concerning Glamdring that keeps it inline
with Orcrist and Sting (how orcs will treat the player when wielding
it).