The Staff of the Archmagi keeps curse protection from Magicbane.
As the commit title implies. And it's odds of defeating a curse are even
better than that of Magicbane.

I also changed the recipe for make the Staff, switched out Vorpal Blade
for Secespita. Felt more appropriate to make the second artifact be more
geared towards a primary caster.

Some code formatting clean up as well in sit.c