New forging artifact recipe - Shadowblade.
Shadowblade is another artifact that can only come into existience via
forging Stormbringer and Grimtooth together. It is an athame that when
wielded, confers stealth and searching, and also casts an aura of
perpetual darkness, the range of which is determined by its BUC status.
It inherets Stormbringers drain life attack and protection, without it
feeling the need to attack peacefuls. In the hands of a rogue,
Shadowblade gives a significant damage bonus to backstabs if wielded in
the primary hand. It can be invoked to cause fear. Is chaotic in
alignment.

Also in this commit - if a wizard forges Glamdring, long sword skill is
unlocked to basic instead of skilled. Forging shouldn't override being
crowned when it comes to unlocking a skill.

There's a bit of code format cleanup here too.