New forging artifact recipe - Ashmar.
Ashmar is a dwarvish roundshield made of mithril, with exceptional
defensive capabilities. It is a forged artifact, and can be created by
forging Trollsbane and Ogresmasher together. Ashmar is attuned to
dwarves, but can be used by any race, and it is non-aligned. Other
stats:

* Ashmar, when worn, provides half-physical damage, hungerless regen,
acid resistance, and will bump con up to 25
* Ashmar can be invoked to heal the wearer (same as the Staff of the
Aesculapius)
* Ashmar also prevents knockback attacks from working
* if Ashmar deflects an attack, there's a 1 in 5 chance of it knocking
back the attacker
* the amount of AC this shield provides is more than the average shield,
especially if the wearer is a dwarf

Thanks krm26 for the initial idea. This is one of those artifacts that
may actually make using weapon/shield more attractive than
dual-wielding.