The Gauntlets of Purity cannot be worn by the impious.
If the players alignment record drops below 20 (pious), they will be
unable to wear the Gauntlets of Purity. If worn, and they do something
to abuse their alignment record, the gauntlets will remove themselves,
and won't be able to be worn again until the player improves their
alignment record. Alignment abuse record does *not* play a factor in how
the gauntlets behave. It's fairly easy to build your alignment record up
to 20 and above - it would take a player consistiently abusing their
alignment to prevent themselves for wearing these regularly.