Fix: elemental vulnerability effects against monsters.
The problem was, if a monster elemental vulnerability spell were used
against another monster, it would modify permonst. That makes it affect
all of its brethren as well, and wouldn't resolve itself until the game
were saved/reloaded. Plus there was no timeout for the effect like there
was for a the player. This commit fixes both issues. Big thanks qt for
the patch.

Co-authored-by: entrez <me@entrez.cc>