Fix: various issues with the Sword of Annihilation's attack.
Couple things - monsters being killed by full on disintegration was
being handled incorrectly. Their inventory was being left behind, but
this is now properly handled. Also noticed that if a monster attacked
another monster with the sword, and a piece of armor was disintegrated,
sometimes the players armor was targeted. That's addressed too.