Fix: proper credit for passive disintegration.
Noticed that in some instances, the player was not receiving credit for
passively disintegrating a monster that was attacking them. This led to
removing passive_disint_mon() and creating disnt_mon_invent(), which
only handles the monsters inventory, and then calling the kill function
outside of it separately (xkilled or monkilled depending on who/what
should earn the credit for the kill).