Fix: impossible, eel hiding out of water (gnomish mines).
The makeriver() routine would run first, placing the rivers down and
populating them with monsters, then remaining functions would run, such
as placing stairs and traps. If the latter happened while an S_EEL type
monster was in that spot, they were still hidden, but on a tile that
they normally wouldn't be hidden on. This commit fixes that. Thanks
hackemslashem for all the fuzzing sessions catching this, and qt for the
insight and patch.

This resolves issue #110.

Co-authored-by: entrez <me@entrez.cc>