Fix: wishing for an object with an invalid material in wizmode while fuzzing.
This addresses a minor annoyance more than anything. If wishing for an
object using an invalid material, that will throw an impossible/program
in disorder, which trips up the fuzzer. The fuzzer likes to do this
semi-often, and I can't recall a time where this impossible has occurred
in normal play. So, change the impossible to a regular pline, only if
fuzzing is enabled.