Fix: monster recharging wands inside of a sack.
Conditions: a monster carrying a sack or other portable container that
had a spent wand inside of it, and also carrying a scroll of charging.
The monster would charge the spent wand without taking it out of the
container first, and if the wand blew up, would cause the game to crash
(extract_from_minvent called on object not in minvent --> dealloc_obj:
obj not free). To fix - make sure that a monster will only recharge a
wand it has out in open inventory. Thanks qt for the nudge in the right
direction.

Possible to-do: have the monster remove a wand from its container if it
wants to recharge it. In current state, it will just leave them inside
the container.