Fix: guard against monster casting a spell at an already dead monster.
Saw this in Hack'EM git commit 3a46159, and took it a bit further. This
guards against you poly'd vs mon as well as mon vs mon, and cleans up
some of the code in each function that seemed overly complicated.

Also fixes a small bug with casting healing on yourself when poly'd into
a spellcasting monster (healing wasn't occurring).