Fix: placement of monster rider/steed when separated.
Long-standing bug regarding monster riders and their steeds, in the
event that one was killed before the other, or if they were somehow
separated and both rider/steed were still alive. Hopefully this greatly
reduces or eliminates the 'program in disorder' errors in-game, placing
a monster on top of another, or the like.

I tested all three scenarios (rider dies but steed lives, steed dies but
rider lives, both rider and steed live but are separated). Could use
further testing.