Allow magic lamps to operate underwater.
I noticed this recently in Gollum's cave - went for a swim there,
because I had acquired a pair of goggles and wanted to search for loot
underwater. Because that level is unlit, going into the water, it was
pitch black (as it should be). Tried turning on a magic lamp, nada,
doesn't work. So I tested out exactly what you can light underwater and
what you can't. Lamps of any kind were a no-go, candles too for obvious
reasons. Magical objects that give off light (sunsword, shield of light,
etc) worked great. So I thought it'd be cool to at least allow magic
lamps to work underwater. With the underwater branch coming, maybe this
warrants creating a new light source, such as a diving lamp that is
immune to water effects or being submerged. For now though, this change
will suffice.