Fix: Sword of Annihilation's disintegration vs target and target's worn armor.
The issues were noticed during a player's game where an NPC had stolen
the sword away from the player and started using it against them. The
player attacking vs the monster attacking wasn't being properly being
passed down through the code - there was a case where the player should
have been fully disintegrated and died, but wasn't, which set off a
cascade of 'program in disorder' messages.

Thanks qt for the assist. This was a confusing bit of code, and this
commit not only fixes the issues observed, but makes it easier to
understand what's going on.

Co-authored-by: entrez <me@entrez.cc>