Fix: Angelslayer's fire damage vs target, Sword of Annihilation's target armor tweaks.
Angelslayer needed the same treatment that the Sword of Annihilation
just received from the commit prior to this one. The way an angelic
being (player or monster) could be killed outright was handled was
incorrect.

Also in this commit - the chance for worn armor to be disintegrated when
attacked by the Sword of Annihilation should occur regardless of whether
the target is disintegration resistant or not. This behavior is inline
with other forms of attack that can destroy/damage objects in open
inventory.