Fix: heap-use-after-free in munstone() function.
Discovered and fixed in Hack'EM, but my solution here is slightly
different. This fix allows any valid object in the monster's inventory
to be affected by stone to flesh on itself.

The next logical step here would be to allow armor and weapons made of
stone to also be affected. Will save that for another day.