Identity of the Riders hidden via farlook or 'More info about' prompt.
Idea is from UnNetHack, but a good bit of this is my own implementation
because of the way the pokedex in this variant works. If trying to use
farlook on Astral Plane to identify one of the riders, it will just
come back with 'Rider' or 'Rider, riding its steed' instead of
displaying their actual name. Same with using the pokedex lookup as a
farlook method. You will be able to ID each Rider by either seeing them
normally from a distance (eyesight) or by being right next to them.