Fix: giant crocodiles and their eggs.
An uncommon scenario - a monster type that can grow up into a larger
monster, which in turn can then grow up into yet another larger monster.
The uncommon bit is that all the grown up monsters in the chain are
oviporous. For crocodiles, it goes like this: baby crocodile ->
crocodile -> giant crocodile. Before this commit, a giant crocodile
would lay a giant crocodile egg, which would hatch a giant crocodile.
It now lays a 'crocodile' egg that hatches as a baby crocodile, as it
should.