Reading a blessed scroll of amnesia allows option to reset accumulated skills.
Inspired by GnollHack's 'scroll of retraining', but here I've decided to
repurpose an existing scroll that typically isn't used for much of
anything, rather than creating an entirely new object.

If the player reads a blessed scroll of amnesia, and their intelligence
score is 13 or greater, they'll have the option to reset all of their
trained up skills, and can reallocate those points differently if they
so choose. You can only put those points back into a skill that had
already been trained up, but the usefulness here is when the player may
have trained up a skill higher than they meant to and are stuck having
those points stay with that skill.

Co-authored-by: entrez <me@entrez.cc>