Fix: prevent Valley of the Dead bones files.
This is considered a fix, because if Cerberus is already dead, and then
the player dies in the valley and they leave bones, the game now becomes
unwinnable for anyone who gets this bones file (have to kill Cerberus to
unlock access to Gehennom, can't do that if he isn't there). Ugh.