Fix: containers flagged as waterproof were not subject to water damage.
An entire block of code was rendered unreachable. Waterproof_container()
encompasses large boxes, chests, and iron safes - all of which can (or
in the case of an iron safe, will) spawn with a material that is
susceptible to water damage. As a reminder, Waterproof_container()
means the contents inside the container are protected from water
damage, not necessarily the container itself.

Move destroyed container handling out of water_damage() and place it
within erode_obj().