Fix: do not allow the Idol of Moloch to be cancelled.
The method in which the game tracks the Idol of Moloch having been
imbued or not is by assigning it a charge, much like a weapon, or a
wand, etc. The flaw in this, is if the Idol is ever cancelled after it's
been imbued, the game becomes unwinnable.

The proper fix for this would be to assign a flag to the Idol, tracking
whether it's been imbued or not. Adding a new flag for this will break
saves, so the solution for now is to just prevent the Idol from being
cancelled altogether. It can still lose it's blessed/cursed status, but
it will still hold on to its 'charge'.