Prevent bypassing demon boss lairs via level teleport.
Now that the player no longer needs to come back up through Gehennom for
the ascension run, it became possible for the player to skip all of the
demon bosses altogether via levelporting past their lairs. You could do
this before Purgatory was introduced, but you still had to come back up
with the Amulet of Yendor/imbued Idol of Moloch, and teleportation is
suppressed at that point. This patch prevents levelporting past the
demon boss lairs. If the player tries to levelport past them, they'll
land on each demon boss level instead. And because using teleportation
on yourself is greatly supressed while that demon boss is alive, it
virtually guarantees each demon boss having to be dealt with.