Fix: the Riders being disintegrated and not leaving a corpse.
This was discovered during a recent ascension where the player was
wielding the Sword of Annihilation, disintegrated a Rider, and they
left no corpse, meaning that Rider would never resurrect. This isn't
supposed to happen, and there is already a special case for this
regarding a black dragon's breath attack. I borrowed a bit of code from
that bit, and added cases for the Riders if disitegrated by the SoA, or
by passive disintegration attacks (black dragon-scaled armor, black
dragon/antimatter vortex). Disintegration gaze attacks (beholders) were
already covered.