Fix: object materials - several wish parsing bugs.
From xNetHack git commit 0ea9e0f, modified. The biggest modified change
was silver/gold dragon scales handling. Original authors notes below:

'Original problem was that I noticed a wish for "gold detection" would
parse out the "gold" and then not know what to do with the rest. This
case itself could have been fixed by adding it to the hardcoded list
that checks for certain things starting with materials, but I decided to
fix the more general problem instead and knock off several other bugs in
the process.

Now, the wish code, when it encounters a material, checks the input
string against all possible things that could be wished for that could
plausibly start with materials: all objects, all object descriptions,
all monsters, all artifacts, all terrain/trap types, effectively
generalizing that hardcoded list to cover all cases. If any of them
match, the material is not treated as a wish for the material. Special
handling was needed for objects where the entire randomized description
is just a material name, such as "silver ring". The good news is that
now when you wish for "silver ring" you actually get the
naturally-silver ring and not a random ring forced into being silver.

As a consequence of making wishes for just "gold" work, I removed
previously existing behavior in which you could wish for just a material
and get an object of that material -- regardless of whether it was valid
for that item to be made of that material. This was never a real feature
anyway. The behavior now appears the same as vanilla, so a wish for just
"silver" will get you a silver ring/spellbook/wand, a wish for "wooden"
will get you the wooden ring, and a wish for "gold" will get you neither
the gold ring, spellbook, or wand, but will return gold pieces.

Checking all objects/monsters/etc is not particularly performant, but
since this only gets called during wish parsing which isn't very
performant anyway, I don't think it's a big deal.'