Regular eggs have a small chance to cancel gnomes.
From 'Tik Tok of Oz', eggs had the power to strip all magical powers
away from gnomes. Cancellation is very powerful magic in NetHack as a
whole, and regular eggs are not uncommon - this power is introduced
here, but only with a small chance of occurring (throwing at gnome
monsters, or eating one as a gnome).