Mind flayer larva victims, if peaceful, will become hostile when turned into a mind flayer.
Old behavior: mind flayer larva attaches itself to its victim, turns it
into a mind flayer. the resulting mind flayer would be peaceful if the
victim was also peaceful. New behavior: resulting mind flayer will be
hostile, regardless of whther the victim was peaceful or not.

If the mind flayer larva was tame, the resulting mind flayer will be
tame. That hasn't changed.