Fix: objects created with an object property would sometimes be blessed and cursed simutaneously.
Odds were low for this to happen, but sometimes it could occur. The
function mksobj() would set the blessed/uncursed/cursed status of an
object, and if it were armor or a weapon, there's a chance
create_oprop() is called. That function also had a chance to bless or
curse the object created, hence the being blessed and cursed at the
same time chance. Odd.