Fix: locking then unlocking the Sokoban closet doors prevents them from being sealed.
Discovered by Janis Papanagnou and reported on the
rec.games.roguelike.nethack discussion group - in Sokoban, the doors to
the prize closets are setup in such a way that once one is opened, the
other two doors disappear and become a vertical wall. If you first
locked and then unlocked a door, that door would no longer be
flagged to disappear and seal up the closet. This allowed physical
access to all three prize closets. Not a huge deal in the grand scheme
of things, as touching one prize will make the other two disappear, but
it does go against the intent of how they're supposed to behave.