Fix: inconsistent attack behavior while submerged in water.
Recently it was discovered that when underwater, the player gains
several advantages while in combat that can make them virtually
invulnerbale to most monsters. This problem exists in vanilla NetHack,
and probably every other variant out there as well. The reason it was
noticed in EvilHack, is because playing as tortle race, you can survive
underwater with no issue at all, and don't need magic or any special
gear to do so.

The problems: player is underwater, a flying/floating/clinging monster
occupies a space next to the player over a water tile, which the player
can see. The player can attack that monster, but their target has no way
to counter attack. Or while underwater and next to dry land, a monster
on dry land next to the player that can swim can attack the player at
will from that spot, but the player can't see where the attack is coming
from and has to guess which direction to attack.

The solutions here do the following, and these all assume that the
player is underwater:

* A monster that can swim has to be in the water to be able to attack
the player, they can't attack from dry land. This gives the player the
ability to attack back without guessing where to attack.
* Flying/floating/clinging monsters that are occupying a water tile can
be seen by the player, but can't be attacked by the player. To prevent
being hemmed in, the player can swim underneath them to move to the next
water tile, much like displacing a tame/peaceful monster.
* The player cannot attack a monster that is on dry land, only those
that are in or under the water.

These changes level the playing field. Basically it now becomes this -
while the player is underwater, only monsters in or under the water can
attack or be attacked by them. An argument could be made that the player
should be able to attack land-locked monsters if adjacent to their
tile and vice versa, since swimming monsters in water can already do
this vs the player. I may revisit this at a later date.

Also in this commit - shield bash and a Monk's extra kick attack are
supressed while attacking underwater.