m_canseeu() allows seeing the player underwater.
Removing !Underwater from m_canseeu() allows monsters that rely on that
function to see the player while submerged, just as the player can see
monsters that are in the water. The one intended side effect of this, is
that several monster spells can now affect the player while they are
submerged. Seems only fair since the player can cast spells against
flying/land-based monsters while the player is underwater.

I didn't check every single instance of m_canseeu(), so it'll be
interesting to see what else this change affects.