Fix: Izchak keeps transforming back into angelic form after every death.
If Izchak the human shopkeeper is killed, he reverts into his 'true
form' as a very powerful archangel. Very difficult to kill, but not
impossible. A change made a couple of years ago to fix a bug with this
process made it so he would transform into an archangel after *every*
death, not just the inital one in human form. This made him
indestructable, much like Kathryn the Ice Queen/Enchantress. This commit
fixes this, so after being killed while in archangel form, he stays
dead. Thanks terrapin for the code fix suggestions.