Fix: player monsters inheriting race flags from the player.
This was an odd one. Was first noticed when a player playing as an
Infidel and was crowned (became a demon) would encounter Infidel player
monsters that were hostile, and they started summoning other player
monster Infidels. The feedback seen matched that of a demon summoning
another demon. The suspicion that race flags were being inherited was
later reinforced when these Infidel player monsters were taking silver
damage, even though their race was something else entirely.

And this wasn't limited to Infidels. Any player monster encountered was
taking on some of the racial properties of the player, like material
hatred.

This commit addresses the summoning bit, material hatred, plus
hit/damage calculations where race is important. There may be other
areas where edits are warranted. I did a search through, and these are
what were apparent. Thanks terrapin and qt for the help with this.