Fix: refactor and fix various bugs in artifact_hit() function.
There were a few things wrong in this function, and it mostly boiled
down to this: several types of effect damage would not occur if you
couldn't see the target of the attack. Reason for this, is because
various effects were wrapped up inside realizes_damage(), and that
relies on being able to see/sense your target. So that's now fixed. Also
did a refactor, removed a fair amount of unnecessary code. The
realizes_damage bits are now inline with how it's done in NetHack 3.7.

Last bit - noncorporeal monsters (ghosts, shades, etc) are immune to
various types of attacks. Before this commit, you could land a hit with
say for example, Frost Brand, but it wouldn't do any damage, which is
misleading. That's been addressed now as well - an uncursed +0 Frost
Brand will pass harmlessly through a ghost, but bless it, or make its
enchantment high enough, it'll land a hit and do regular physical
damage.