Make the archangel Saint Michael a unique monster.
The named archangel Saint Michael at the end of Purgatory has been
upgraded, and is now a unique angel. He retains many of the properties
of a normal archangel, but can also employ the magic missile attack of a
regular angel.

I may make some minor adjustments to this, but this is a good start.
While not as strong as the demon princes encountered in Gehennom, Saint
Michael is definitely more powerful than his earlier form as a regular
archangel.