Create new artifact flag and macro (SPFX_FORGED and is_forged_artifact()).
I was initially making this to ensure that forged artifacts would never
appear in a bones level. Turns out that this is already handled, because
currently all forged artifacts are also non-wishable. Didn't want to
waste the work done, so this can stay in, as the macro may become useful
down the road.