Gauntlets of Purity tweaks.
If playing as an Infidel, you can't wear the Gauntlets of Purity at all
(using a helm of opposite alignment won't help in this case). If
chaotic (or somehow become unaligned and not an Infidel) and are able
to wear the gauntlets, there's a very small chance that the artifact
will sense your misalignment and remove themselves randomly as time
passes (they could then be reworn if still pious). If your alignment
does not match those of the gauntlets (read: not lawful), the damage
dealt from the blast from touching the gauntlets is signifcantly more
than normal.