Fix: more separate_steed_and_rider() tweaks, other monster/steed tweaks.
This is a big one, this fix gets rid of many 'program in disorder'
errors many players have witnessed before. In the function
separate_steed_and_rider(), some extra handling was needed when a
monsters steed died, but not the rider. I also corrected a bit of code
that would determine when a monster would try to mount a monster to ride
it (Rodney isn't allowed to have steeds anymore, so no need to check for
his chances of doing so).

One last thing - a bit of changes to mfind_target, suspect that
mm_aggression being limited to the player may have caused a couple
crashes in the recent past. Some formatting clean up too.