New forging artifact recipe - Hammer of the Gods.
Another new forged artifact, known as the Hammer of the Gods. It is
created by forging Sunsword and Demonbane together. The end result is a
silver heavy war hammer that virtually has the same abilities as both
Sunsword and Demonbane - will do special damage to both demons and
undead, warns of both (and will make both angry/untame when wielded),
gives off light when wielded, and will suppress demon gating.

The extra bits added to this artifact that sets it apart from the
artifacts used to create it - will grant sickness resistance when
wielded, and it can be tethered and thrown like an aklys if the players
skill in using a hammer is at skilled or better.