Fix: monsters stuck in a loop switching between melee and ranged weapon.
When calculating distances between the player and a monster, the value
chosen looks like this:

.........
....9....
..85458..
..52125..
.941@149.
..52125..
..85458..
....9....
.........

With @ being the player. The problem was that one value was chosen for
when the monster would pick a weapon (less than or equal to 8) and that
was different than the value for when the monster would decide whether
to use a melee or ranged (greater than 4). This made position 8 the
'sweet spot' where the monster would get confused and get stuck in a
loop switching back and forth endlessly between ranged and melee
weapons. The fix was to change 'greater than 4' to 'greater than 8' for
deciding when to switch to ranged attack.

This all comes from SpliceHack, which got it from GruntHack and cleaned
it up a bit. It could still use some work undoubtedly, but overall I
still prefer this monster AI behavior vs vanilla NetHack. There's some
code formatting cleanup in this commit as well.