Fix: monster steed being knocked back while being ridden.
There's a couple fixes in this commit, but the big one is this:
mhurtle() could handle a monster rider being knocked off of its steed
just fine, but there wasn't any specific handling for a monster steed
being knocked back instead. This could happen if the monster steed were
attacked by another monster that used AD_CLOB, of if the monster steed
attacked a player wearing Ashmar. before this commit, the knockback
would still happen, but very weird and bad things would happen, program
in disorder messages would fly by, and then a crash from interacting
with 'tail of a <monster>, being ridden'. So this long-standing bug is
now fixed. The monster steed is knocked back, and it's properly
separated from its rider.

Also in this commit - in a couple places where mhurtle() was called,
there was no guard aginst whether the monster being hurtled was dead or
not. Ashmar could also still try to hurtle an attacker if they had died
via the act of touching it (material hatred). Those instances are fixed
as well.

Last bit - relabeled a couple pointers for riders/steeds so they'd make
more sense at first glance and not be so misleading. rider_id becomes
ridden_by, and m1 becomes mon_steed.

Thanks krm26, terrapin, and qt for the help on this.