New forging artifact recipe - Tempest.
Yet another new forged artifact, known as Tempest. It is created by
forging Cleaver and Mjollnir together. What comes out of the forge is an
iron halberd that has the same lightning attack and protection (with a
better to-hit) as Mjollnir, but it does not inherit the multi-target
attack from Cleaver. Instead, there's a 20% chance that an extra
lightning area of effect attack will occur, affecting the target and
anything right next to it.

Due to the nature of how polearm-type weapons do damage, none of the
lightning-based effects will happen if the target being attacked (player
vs monster) is within close melee range; only via ranged targeted
attack. There are only two exceptions to this - using Tempest while
riding a steed, or playing as Centaur race.

Tempest also grants immunity to being stunned when wielded, and can be
invoked for conflict.

This could be tweaked further. There's a minor feedback/display bug when
a monster attacks the player with this artifact. I may revisit the area
of effect attack as well, being caught in a ball of lightning is not
good for rings and wands in open inventory, which would be a deterrent
to using this.