Livelog wishes granted exclusively by the Magic 8-Ball.
Because the Magic 8-ball only has a 1 in 100,000 chance of granting a
wish when applied, adjust the livelog so it states exactly where the
wish came from. This could probably be expanded upon to livelog each
specific source of the wish.

Also fixed a minor feedback bug - since the 8-Ball goes away after
granting a wish, it never gave one last rumor. Was probably never
noticed, because as far as I know, no one has ever gotten a wish from
the 8-Ball.