Wielding or wearing an object that grants petrification resistance will cure it if in the process of being petrified.
In EvilHack, the only objects that grant stoning resistance are the
Sword of Kas, and yellow-dragon scaled armor. Trying to put on
yellow-dragon scaled armor really doesn't help if being stoned, because
it takes too many turns to put on body armor, even if very fast. But
it's included here because it's the correct thing to do.