Fix: objects being created with materials they shouldn't have.
The main fix here is the 'ashwood staff', it's the base item for the
Staff of the Archmagi. It doesn't generate randomly, but a player could
wish for it, and wind up with an 'iron ashwood staff'.

Also made it so mummy wrappings keep their base material. Not really
appropriate to see one made of leather or dragonhide.