New forging artifact recipe - Ithilmar.
Ithilmar is an forged artifact barding that is created by forging Vorpal
Blade and the Gauntlets of Purity together. When worn by a steed, it
confers magic resistance and reflection, and is also spiky, meaning that
any monster that tries to engulf a steed wearing this artifact will be
spit back out immediately. It's material is steel, weighs less than
other types of barding, and gives an extra three points of AC to the
steed wearing it vs other types of barding. It can be enchanted just
like any other type of barding.

Ithilmar has some other unique effects that makes it the ideal gear for
any steed the player wishes to keep alive:

* Ithilmar inherits Vorpal Blade's decapitation protection and imparts
that to the steed
* Jousting with a steed wearing this barding makes the lance used immune
to breakage
* A steed wearing Ithilmar can be left alone/abandoned indefintely;
their tameness will not decrease
* Also while wearing Ithilmar, the steed will never starve. They will
still have a period of time where they become hungry and will want to
eat food, but starvation will never occur
* A steed wearing Ithilmar will always resist the effects of conflict
* While wearing Ithilmar, a steed will not lose tameness when being
mounted

This is the ultimate piece of gear for the player's steed, but it comes
at a cost of four total artifacts - the two used to make the Gauntlets
of Purity, the gauntlets themselves, and Vorpal Blade. If you're the
type of player that values their steed and would like to have them
ascend with you, then the cost is worth it for this forged artifact.

There may be some more tweaks to this in the future, but right now this
feels like it's dialed in pretty well.