New forging artifact recipe - Armor of Retribution.
This may be the final forged artifact recipe I create, and this may be
one of, if not the most powerful forged artifact of them all. The Armor
of Retribution is crystal plate mail that is created by forging two
forged artifacts together - Ashmar and the Sword of Annihilation. This
means that a total of six artifacts will be used up to create the one,
but the abilities of this body armor may be worth it to some. The stats:

* The Armor of Retribution (AoR) enjoys all of the special properties of
being crystal plate, meaning it doesn't inhibit spellcasting, is immune
to being disintegrated or subject to a monsters 'destroy armor' spell,
and so forth
* The AoR inherits many things from the artifacts used to forge it,
including imparting disintegration resistance (worn), acid resistance
(carried), half physical damage, makes the wearer immune to knockback
attacks, and can sometimes knockback an attacker if the armor blocked
the physical attack. It is also an indestructible object, gets an AC
boost vs regular crystal plate mail, and grants 25 constitution when
worn
* Other abilities gained: half spell damage, the ability to
sometimes protect the wearer against cancellation, and since this
artifact is considered indestructible, it won't break apart if the
player wearing it polymorphs into something too large to wear it (it
falls off instead)
* The AoR cannot be dragon-scaled. The attempt will fail, and will use
up the scroll of enchant armor used in the attempt

One other unrelated change in this commit - I noticed an area where
Ashmar should have protected against knockback and didn't, that was
included when adding the AoR.

Ashmar and the Sword of Annihilation are very powerful artifacts in
their own right, some players may find those difficult to give up to
make the Armor of Retribution. Given how powerful this body armor is, I
think the sacrifice may be worth the cost. We'll see, it'll be
interesting to see what strategies players come up with.